refactor(core) :  interface, dto 삭제
