--- src/main/java/calculator/input/InputDto.java
@@ -1,13 +1,0 @@
-package calculator.input;
-
-public class InputDto {
-    private String input;
-
-    public InputDto(String input) {
-        this.input = input;
-    }
-
-    public String getInput() {
-        return input;
-    }
-}
